Image_classification/requirements.txt
tensorflow>=2.13.0
streamlit>=1.32.0
pillow>=10.0.0
numpy>=1.24.0
